Fix issue with event parameters in squirt day helper having wrong contents
The event handlers in the squirt day helper assumed that the event name is given
as first parameter (as is usual with the OnEvent-script), although it is being
removed by the OnEvent-script in `src/base.lua`. Since said script calls an
event handler based on the event name, the event name is always the same in a
given event handler.