Remove realm and faction-specific setting for item mailer character
With the recent warband and cross-faction changes we should be able to mail items
to any character regardless of faction and realm. There also were errors being caused
by calling the function UnitFactionGroup before the event PLAYER_ENTERING_WORLD has
fired, which happened on the initial population of the settings frames due to the
faction and realm-specific setting.